+++ src/browserbehavior/BrowserBehavior.ts
@@ -96,4 +96,14 @@ export default interface BrowserBehavior {
    * Returns whether browser supports setSinkId operation
    */
   supportsSetSinkId(): boolean;
+
+  /**
+   * Returns whether browser supports the playback of canvas-captured stream.
+   */
+  supportsCanvasCapturedStreamPlayback(): boolean;
+
+  /**
+   * Returns whether browser supports video processing based on tensforflow models.
+   */
+  supportsVideoProcessingBasedOnTensorflowModel(): boolean;
 }