+++ CHANGELOG.md
@@ -4,10 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
-
-## [2.2.1] - 2020-12-11
+## [Unreleased]
 
 ### Added
+- Add `DefaultVideoTransformDevice` to implement `VideoTransformDevice`.
+ `VideoFrameProcessor`, `VideoFrameProcessorPipeline` and `VideoFrameBuffer` interfaces
+ are added to support `DefaultVideoTransformDevice` and allows processing steps to be applied to device.
+ The method `chooseVideoInputDevice` in `DefaultDeviceController` can handle `VideoTransformDevice` now.
+
+
+## [2.2.1] - 2020-12-11
 
 - [Demo] Fatal errors in the demo will now cause a fatal screen if `fatal=1` is present in the
   URL, or if the demo is running locally.